feat: Add support for spheres in raytracing

+++ include/utils.h
@@ -6,7 +6,7 @@
 /*   By: descamil <descamil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/03 12:37:54 by descamil          #+#    #+#             */
-/*   Updated: 2024/12/19 11:36:50 by descamil         ###   ########.fr       */
+/*   Updated: 2025/01/23 12:53:28 by descamil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include "../minirt.h"
 
 // ft_atof.c
-float	ft_atof(char *str);
+double	ft_atof(char *str);
 
 // ft_atoi_mod.c
 int		ft_atoi_mod(char *str, int *error);
